requirements.txt
# Core GUI Framework
customtkinter>=5.2.1
Pillow>=10.3.0

# System Monitoring
psutil>=5.9.8

# Network Tools
speedtest-cli>=2.1.3

# Configuration Management
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# Testing
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Security
cryptography>=42.0.4

# Scheduling
schedule>=1.2.1

# Report Generation
reportlab>=4.0.9
jinja2>=3.1.3

# Type Checking (Development)
mypy>=1.8.0

# Code Quality (Development)
black>=24.1.1
flake8>=7.0.0
pylint>=3.0.3
